+++ openBCISimulator.js
@@ -3,8 +3,8 @@ const EventEmitter = require('events').EventEmitter;
 const util = require('util');
 const stream = require('stream');
 
-const obciUtilities = require('openbci-utilities').Utilities;
-const k = require('openbci-utilities').Constants;
+const obciUtilities = require('openbci-utilities/dist/utilities');
+const k = require('openbci-utilities/dist/constants');
 const now = require('performance-now');
 const Buffer = require('safe-buffer').Buffer;
 
@@ -212,7 +212,7 @@ Simulator.prototype.write = function (data, callback) {
     case k.OBCIMiscSoftReset:
       if (this.stream) clearInterval(this.stream);
       this.streaming = false;
-      this._output(new Buffer(`OpenBCI V3 Simulator On Board ADS1299 Device ID: 0x12345 ${this.options.daisy ? `On Daisy ADS1299 Device ID: 0xFFFFF\n` : ``} LIS3DH Device ID: 0x38422 ${this.options.firmwareVersion === k.OBCIFirmwareV2 ? `Firmware: v2.0.0\n` : ``}$$$`));
+      this._output(new Buffer(`OpenBCI V3 Simulator On Board ADS1299 Device ID: 0x3E ${this.options.daisy ? `On Daisy ADS1299 Device ID: 0x3E\n` : ``} LIS3DH Device ID: 0x38422 ${this.options.firmwareVersion === k.OBCIFirmwareV2 ? `Firmware: v2.0.0\n` : ``}$$$`));
       break;
     case k.OBCISDLogForHour1:
     case k.OBCISDLogForHour2: